+++ CHANGELOG.md
@@ -5,12 +5,21 @@ All notable changes to [**TNSCM** *(Tenable Nessus CLI Manager)* by LimberDuck][
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.2] - 2021-08-25
+
+### Added
+
+- `plugin --family-list` lists parameters `id`, `name`, `count`
+
+### Changed
+
+- date format for returned dates
 
 ## [0.0.1] - 2021-08-24
 
 - initial release
 
-
+[0.0.2]: https://github.com/LimberDuck/tnscm/compare/v0.0.1...v0.0.2
 [0.0.1]: https://github.com/LimberDuck/tnscm/releases/tag/v0.0.1
 
 [1]: https://github.com/LimberDuck/tnscm